contrib/assets/SwitchButton: rename to CheckboxButton
